Solve circular imports for results
