+++ src/matchbox/client/models/models.py
@@ -5,12 +5,12 @@ from pandas import DataFrame
 
 from matchbox.client.models.dedupers.base import Deduper
 from matchbox.client.models.linkers.base import Linker
-from matchbox.client.results import (
+from matchbox.common.exceptions import MatchboxResolutionNotFoundError
+from matchbox.common.results import (
     ModelMetadata,
     ModelType,
     Results,
 )
-from matchbox.common.exceptions import MatchboxResolutionNotFoundError
 from matchbox.server import MatchboxDBAdapter, inject_backend
 from matchbox.server.base import MatchboxModelAdapter
