+++ src/matchbox/server/postgresql/adapter.py
@@ -5,7 +5,6 @@ from pydantic import BaseModel
 from sqlalchemy import and_, bindparam, delete, func, or_, select
 from sqlalchemy.orm import Session
 
-from matchbox.client.results import Results
 from matchbox.common.exceptions import (
     MatchboxDataNotFound,
     MatchboxResolutionNotFoundError,
@@ -36,10 +35,14 @@ if TYPE_CHECKING:
     from pandas import DataFrame as PandasDataFrame
     from polars import DataFrame as PolarsDataFrame
     from pyarrow import Table as ArrowTable
+
+    from matchbox.common.results import Results
+
 else:
     PandasDataFrame = Any
     PolarsDataFrame = Any
     ArrowTable = Any
+    Results = Any
 
 
 class FilteredClusters(BaseModel):
