+++ test/client/test_linkers.py
@@ -10,6 +10,7 @@ from matchbox.client.models.linkers.splinklinker import (
     SplinkLinkerFunction,
     SplinkSettings,
 )
+from matchbox.client.results import results_to_matchbox
 from matchbox.common.sources import Source
 from matchbox.server.postgresql import MatchboxPostgres
 
@@ -174,7 +175,7 @@ def test_linkers(
 
     # 4. Probabilities and clusters are inserted correctly
 
-    results.to_matchbox(backend=matchbox_postgres)
+    results_to_matchbox(results=results)
 
     model = matchbox_postgres.get_model(model=linker_name)
     assert model.results.probabilities.shape[0] == fx_data.tgt_prob_n
