+++ test/fixtures/db.py
@@ -11,6 +11,7 @@ from sqlalchemy import Engine, create_engine
 from sqlalchemy import text as sqltext
 
 from matchbox import index, make_model
+from matchbox.client.results import results_to_matchbox
 from matchbox.common.sources import Source, SourceAddress
 from matchbox.server.base import MatchboxDatastoreSettings, MatchboxDBAdapter
 from matchbox.server.postgresql import MatchboxPostgres, MatchboxPostgresSettings
@@ -103,7 +104,7 @@ def db_add_dedupe_models_and_data() -> AddDedupeModelsAndDataCallable:
                 )
 
                 results = model.run()
-                results.to_matchbox(backend=backend)
+                results_to_matchbox(results=results)
                 model.truth = 0.0
 
     return _db_add_dedupe_models_and_data
@@ -173,7 +174,7 @@ def db_add_link_models_and_data() -> AddLinkModelsAndDataCallable:
                 )
 
                 results = model.run()
-                results.to_matchbox(backend=backend)
+                results_to_matchbox(results=results)
                 model.truth = 0.0
 
     return _db_add_link_models_and_data
